connmgr: Drop 'next_in_band_update' timer.
This timer used to exist because it was possible for in-band remote IP
addresses to change without any intentional configuration change in one
case: where controller discovery found a new controller.  Controller
discovery was removed long ago, but the reason for the timer had been
forgotten and so remained.  This commit removes it.