ChangeLog: Rename NEWS.
Many users seem to miss changes that Open vSwitch clearly states in
ChangeLog.  Some of this is probably due to users not reading any
documentation ever, but I wonder whether some of it is because users
expect ChangeLog to be akin to a commit log, because that's what GNU
software puts in ChangeLog.  Users don't read GNU ChangeLogs, because
they don't contain information useful to users.

GNU software puts what we have in ChangeLog in a file named NEWS.  Some
users read it.  Therefore, this commit renames ChangeLog to NEWS.

Also, strip trailing whitespace for Ethan-compliance.