ofproto-dpif: Make ofproto/trace accept an odp_flow in place of a packet.
It's often easier to get a flow than it is to get a packet.  For example,
"ovs-dpctl dump-flows" prints flows, but it doesn't print any of the
packets in those flows.  It is also easier to construct a flow by hand than
it is to construct packet bytes.

This commit, therefore, makes ofproto/trace accept a flow specification in
place of packet data.  In a few cases a flow by itself is not sufficient
to determine what would happen to a packet, so in those cases the command
prints a message to that effect.

An upcoming commit will also start using this feature to unit-test action
translation in ofproto-dpif.

Suggested-by: Reid Price <reid@nicira.com>